refactor(team): SDK GlobalExceptionHandler와 충돌 방지를 위해 예외 핸들러 분리
